+++ src/frontend/apps/drive/src/features/i18n/conf.ts
@@ -1,3 +1,2 @@
 export const LANGUAGES_ALLOWED = ["en-us", "fr-fr"];
 export const LANGUAGE_LOCAL_STORAGE = "main-language";
-export const BASE_LANGUAGE = "fr-fr";